requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
openai==1.3.7
google-api-python-client==2.108.0
anthropic==0.7.8
pydantic==2.5.0
pydantic-settings==2.1.0
importlib-metadata==6.8.0
typing-extensions==4.8.0
asyncio==3.4.3
aiofiles==23.2.1
python-multipart==0.0.6
jinja2==3.1.2

# Neue Dependencies für KI-Projektmanagement-System
streamlit==1.29.0
tree-sitter==0.20.4
esprima==4.0.1
astroid==3.0.1
ollama==0.1.6
pymongo==4.6.0
redis==5.0.1
celery==5.3.4
safety==2.3.5
radon==6.0.1
lizard==1.17.10
psutil==5.9.6

# New dependencies for HumanLayer integration and enhanced features
humanlayer>=0.7.0
mongoengine>=0.27.0
plotly>=5.18.0
streamlit-ace>=0.1.1

# Type checking and code quality
mypy==1.8.0
types-requests==2.31.0.20240106
types-PyYAML==6.0.12.12
pytest-cov==4.1.0
black==23.12.1
isort==5.13.2
flake8==7.0.0